integration du jeu nodebuster

+++ Script/compilation.sh
@@ -10,10 +10,15 @@ cd projet
 #PENSER A REMETTRE COMPILATION JEUX!!!
 for i in *
 do
-    cd $i
-    echo "Compilation du jeu "$i
-    echo "Veuillez patienter"
-    javac -cp .:../..:/home/pi/git/MG2D *.java
+    cd "$i"
+    echo "Compilation du jeu $i"
+    if [ -f pom.xml ]; then
+        echo "  -> projet Maven détecté, utilisation de mvn package"
+        mvn clean package -q
+    else
+        echo "  -> compilation Java basique"
+        javac -cp .:../..:/home/pi/git/MG2D *.java
+    fi
     cd ..
 done
